Add B000FTFT (as well as B000FTF) to the list of layers we should ignore when constructing the shortlink.
The shortlink was also broken for the new /?{node,way,relation}=id
construct since the code assumed that all URL parameters that weren't
lat/lon/zoom/layers could be safely ignored. I've added
node/way/relation to that list but for future-proofness it might be
better to simply check if there's anything left in ''args'' that we
don't know how to encode, rather than checking on values we know about
now.


git-svn-id: http://svn.openstreetmap.org/sites/rails_port@16216 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4