Potlatch's i18n support predates our Rails i18n effort, so it was still doing its own language detection.
Scrap that and use I18n.locale instead, now Potlatch is guaranteed to
use whatever translation the rest of the site uses.

I've tested this by switching the language in preferences back and
forth, and also ensured that Potlatch will fall back on English if the
specified language doesn't exist in its list of translations.


git-svn-id: http://svn.openstreetmap.org/sites/rails_port@16506 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4